In AmlSpec.rtf, change hanging indents
so that wrapped lines are more readable

For lines that are tables with the second column starting at 3 inches, I made the hanging indent 3.25 inches. For other lines, I resized my window to a somewhat small size and looked for lines that wrap. When they wrapped, I manually changed the hanging indent to a level that made sense for that particular line of code. For example, long comments that start with `; this is …` would wrap to right below the ‘t’ in “this”.

I also changed the hyphen in “AML - A Modular Language” at the beginning to a dash “–”, to have a better first impression.